+++ packages/components/src/common/dateUtils/getMidnight/getMidnight.js
@@ -1,4 +1,0 @@
-export function getMidnight(date) {
-  if (!date) return null;
-  return new Date(date.getFullYear(), date.getMonth(), date.getDate());
-}