+++ packages/components/src/common/dateUtils/index.js
@@ -1,5 +1,4 @@
 export { getDayNames } from './getDayNames';
-export { getMidnight } from './getMidnight';
 export { moveToWithinRange } from './moveToWithinRange';
 export { isWithinRange } from './isWithinRange';
 export { getMonthNames } from './getMonthNames';